+++ nginx-lua-prometheus-0.20200420-1.rockspec
@@ -21,6 +21,8 @@ build = {
     type = "builtin",
     modules = {
         ["prometheus"] = "prometheus.lua",
+        ["prometheus_keys"] = 'prometheus_keys.lua'
+        ["prometheus_lock"] = 'prometheus_lock.lua'
         ["prometheus_resty_counter"] = 'prometheus_resty_counter.lua'
     }
 }
