+++ prometheus_lock.lua
